# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auto renginiai API",
        "description": "Auto renginių, registracijų ir autentifikacijos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Lokali aplinka"
        }
    ],
    "paths": {
        "/api/registruotis": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Vartotojo registracija ir tokeno gavimas",
                "operationId": "2bf7b2ba876d503547ba0205a0df37aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vardas",
                                    "el_pastas",
                                    "slaptazodis"
                                ],
                                "properties": {
                                    "vardas": {
                                        "type": "string",
                                        "example": "Lukas"
                                    },
                                    "el_pastas": {
                                        "type": "string",
                                        "example": "lukas@auto.lt"
                                    },
                                    "slaptazodis": {
                                        "type": "string",
                                        "example": "slaptazodis"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sėkminga registracija",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "zinute": {
                                            "type": "string",
                                            "example": "Registracija sėkminga"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validacijos klaida"
                    }
                }
            }
        },
        "/api/prisijungti": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Prisijungimas ir tokeno gavimas",
                "operationId": "c576bc4f567ee44ca10330b575c6c9ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "el_pastas",
                                    "slaptazodis"
                                ],
                                "properties": {
                                    "el_pastas": {
                                        "type": "string",
                                        "example": "org@auto.lt"
                                    },
                                    "slaptazodis": {
                                        "type": "string",
                                        "example": "slaptazodis"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sėkmingas prisijungimas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "zinute": {
                                            "type": "string",
                                            "example": "Prisijungta"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validacijos klaida"
                    }
                }
            }
        },
        "/api/atsijungti": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Atsijungti (ištrina current token)",
                "operationId": "998d592def054d679ebf24d8740f1128",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "zinute": {
                                            "type": "string",
                                            "example": "Atsijungta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neprisijungęs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/as": {
            "get": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Gauti prisijungusio vartotojo informaciją",
                "operationId": "e04560b13175af182f5fa832f85de6c8",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Neprisijungęs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}